Ajustar Makefiles e dependencias
