+++ configure.ac
@@ -4,7 +4,7 @@
 AC_PREREQ([2.69])
 AC_INIT([LivreNFE], [0.1], [pggllrd@gmail.com])
 AM_INIT_AUTOMAKE([livrenfe], [2.6])
-AC_OUTPUT([Makefile src/Makefile])
+AC_OUTPUT([Makefile src/Makefile include/Makefile src/libnfe/Makefile include/libnfe/Makefile])
 AC_CONFIG_HEADERS([config.h])
 AM_INIT_AUTOMAKE([subdir-objects])
 AC_CONFIG_SRCDIR(src)
