+++ src/Makefile.am
@@ -10,13 +10,13 @@ livrenfe_SOURCES = main.c db.c db_create.c db_init.c livrenfe.c lnfe_window.c \
 		   resources.c  db_nfe.c nfe_manager.c item_manager.c \
 		   crypto_interface.c \
 		   db_prefs.c emitente_manager.c sefaz_response.c \
-		   prefs.c gtk_common.c
+		   prefs.c prefs_dialog.c gtk_common.c
 
-livrenfe_LDADD = libnfe.la
+livrenfe_LDADD = libnfe/libnfe.la
 
 resources.c: livrenfe.gresource.xml window.ui nfe_manager.ui item_manager.ui \
 		emitente_manager.ui sefaz_response.ui icons/livrenfe.png \
-		icons/livrenfe_128x128.png prefs.ui
+		icons/livrenfe_128x128.png prefs_dialog.ui
 	glib-compile-resources --target=$@ --generate-source $<
 
 clean-local:
