+++ src/db_prefs.c
@@ -18,6 +18,7 @@
 
 #include "db_interface.h"
 #include "db.h"
+#include "livrenfe.h"
 #include <libnfe/errno.h>
 #include <sqlite3.h>
 #include <string.h>
