+++ src/libnfe/Makefile.am
@@ -3,7 +3,7 @@ lib_LTLIBRARIES = libnfe.la
 libnfe_la_SOURCES = gen_xml.c nfe.c sefaz.c send.c \
 		    sign.c utils.c xml.c
 
-libnfe_la_LDFLAGS = -L/usr/local/lib -Llibnfe -lnfe -lssl -lcrypto -lp11 \
+libnfe_la_LDFLAGS = -L/usr/local/lib -lssl -lcrypto -lp11 \
 	`xml2-config --libs` `pkg-config --libs xmlsec1-openssl`\
        	`curl-config --libs`
