+++ src/livrenfe.gresource.xml
@@ -6,7 +6,7 @@
     <file preprocess="xml-stripblanks">item_manager.ui</file>
     <file preprocess="xml-stripblanks">emitente_manager.ui</file>
     <file preprocess="xml-stripblanks">sefaz_response.ui</file>
-    <file preprocess="xml-stripblanks">prefs.ui</file>
+    <file preprocess="xml-stripblanks">prefs_dialog.ui</file>
     <file>icons/livrenfe.png</file>
     <file>icons/livrenfe_128x128.png</file>
   </gresource>
